chat_service/requirements.txt
fastapi
uvicorn
kafka-python
boto3
psycopg2-binary
python-dotenv
pydantic